requirements.txt
dj-database-url==3.0.1
Django==5.2.4
mysqlclient==2.2.7
djangorestframework==3.16.0
rest_framework_simplejwt==0.0.2